pyproject.toml
[tool.poetry]
name = "repl_python3_RockMeGOD"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
discord = "^1.0.1"
dblpy = "^0.4.0"
qrcode = "^6.1"
dhooks = "^1.1.4"
Pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
